+++ CHANGELOG.md
@@ -5,6 +5,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.6.1] - 2021-04-09
+### Removed
+- Removed order status check and reopening a cancelled order process from the Success controller, since it is already being done by the Notification controller instead.
+
+## [1.6.0] - 2021-03-26
+### Removed
+- Removed obsolete emandate field from Direct Debit checkout
+
+### Fixed
+- Fixed a bug where the wrong processing status was set for the notifications
+
+### Changed
+- Refactored notification controller and improved logging
+
 ## [1.5.0] - 2021-03-11
 ### Added
 - Added generic gateways and generic giftcards to the checkout dynamically.